Refleciton 5/8

+++ Source/Runtime/Platform/PlatformType.h
@@ -1,11 +1,12 @@
 #pragma once
 #include <Runtime/Core/Core.h>
 #include <Runtime/Reflection/Reflection.h>
+#include "PlatformType.reflected.h"
 
 namespace Portakal
 {
 	PENUM();
-	enum class RUNTIME_API PlatformType : byte
+	enum class RUNTIME_API PlatformType : int64
 	{
 		Windows = 112,
 		Linux,
@@ -14,3 +15,39 @@ namespace Portakal
 		Switch
 	};
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
